+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed the potential for a race condition with `cached_path()` when extracting archives. Although the race condition
   is still possible if used with `force_extract=True`.
 - Fixed `wandb` callback to work in distributed training.
+- Fixed `tqdm` logging into multiple files with `allennlp-optuna`.
 
 
 ## [v2.4.0](https://github.com/allenai/allennlp/releases/tag/v2.4.0) - 2021-04-22